renamed files  names with components
