+++ week1/project/ecommerce/src/components/category.jsx
@@ -1,6 +1,6 @@
-function Category({ name, onClick }) {
+function Category({ name, onClick, className }) {
   return (
-    <button className="category-btn" onClick={onClick}>
+    <button className={className} onClick={onClick} >
       {name}
     </button>
   );
