+++ week1/project/ecommerce/src/components/product.css
@@ -2,23 +2,25 @@
   display: flex;
   flex-wrap: wrap;
   list-style: none;
-  margin-block-end: 0;
+  /* margin-block-end: 0;
   margin-block-start: 0;
-  padding-inline-start: 0;
+  padding-inline-start: 0; */
 }
 .product-item {
-  display: list-item;
+  /* display: list-item; */
 }
 .product-item-block {
   display: flex;
   flex-direction: column;
   height: 100%;
   max-width: 300px;
+  box-shadow: 0 2px 4px rgb(0 0 0 / 10%);
+  margin: 5px;
 }
 .product-img {
   margin-bottom: auto;
   margin-top: auto;
-  height: 200px;
+  height: 100%;
   object-fit: contain;
   width: 100%;
   padding: 3px;
@@ -33,5 +35,5 @@
   width: 100%;
   padding: 5px 0;
   height: 60px;
-  box-sizing: border-box;
+  /* box-sizing: border-box; */
 }
